requirements.txt
pytest==8.3.5
requests==2.32.3
python-dotenv==1.0.1
pytest-html==4.1.1
